# Faster region iterators