--- a/src/autopas/containers/CellBlock3D.h
+++ b/src/autopas/containers/CellBlock3D.h
@@ -392,13 +392,13 @@ inline ParticleCell &CellBlock3D<ParticleCell>::getCell(const std::array<index_t
 template <class ParticleCell>
 inline std::array<typename CellBlock3D<ParticleCell>::index_t, 3> CellBlock3D<ParticleCell>::index3D(
     index_t index1d) const {
-  return ThreeDimensionalMapping::oneToThreeD(index1d, _cellsPerDimensionWithHalo);
+  return utils::ThreeDimensionalMapping::oneToThreeD(index1d, _cellsPerDimensionWithHalo);
 }
 
 template <class ParticleCell>
 inline typename CellBlock3D<ParticleCell>::index_t CellBlock3D<ParticleCell>::index1D(
     const std::array<index_t, 3> &index3d) const {
-  return ThreeDimensionalMapping::threeToOneD(index3d, _cellsPerDimensionWithHalo);
+  return utils::ThreeDimensionalMapping::threeToOneD(index3d, _cellsPerDimensionWithHalo);
 }
 
 template <class ParticleCell>