--- a/src/autopas/containers/CellBorderAndFlagManager.h
+++ b/src/autopas/containers/CellBorderAndFlagManager.h
@@ -8,30 +8,30 @@
 
 namespace autopas {
 /**
- * interface class to handle cell borders and cell types of cells.
+ * Interface class to handle cell borders and cell types of cells.
  * @todo: add cell border handling
  */
 class CellBorderAndFlagManager {
   /**
-   * the index type to access the particle cells
+   * The index type to access the particle cells.
    */
   typedef std::size_t index_t;
 
  public:
   /**
-   * destructor
+   * Cestructor
    */
   virtual ~CellBorderAndFlagManager() = default;
 
   /**
-   * checks if the cell with the one-dimensional index index1d is a halocell.
+   * Checks if the cell with the one-dimensional index index1d is a halocell.
    * @param index1d the one-dimensional index of the cell that should be checked
    * @return true if the cell is a halo cell
    */
   virtual bool isHaloCell(index_t index1d) const = 0;
 
   /**
-   * checks if the cell with the one-dimensional index index1d is owned by this
+   * Checks if the cell with the one-dimensional index index1d is owned by this
    * process, i.e. it is NOT a halo cell.
    * @param index1d the one-dimensional index of the cell that should be checked
    * @return true if the cell is owned by the current process.