… examples/sph-mpi/sph-main-mpi.cpp
… examples/sph/sph-main.cpp
… src/autopas/containers/cellPairTraversals/SlicedTraversal.h
… src/autopas/iterators/ParticleIterator.h
… src/autopas/iterators/RegionParticleIterator.h
… src/autopas/particles/Particle.h
… src/autopas/utils/inBox.h
… tests/testAutopas/InBoxTest.cpp
… tests/testAutopas/RegionParticleIteratorTest.cpp
… tests/testAutopas/RegionParticleIteratorTest.h
… tests/testAutopas/testingHelpers/GaussianGenerator.h
… tests/testAutopas/testingHelpers/RandomGenerator.h