The region iterator now...
- [x] starts directly at the region of interest.
- [x] stops directly after the region of interest.
- [x] jumps over uninteresting parts of the cell vector instead of testing all cells.
- [x] test for regioniterator in direct sum
- [x] works with >1 thread
- [x] clean up region iterator tests

This PR aims to close #2 

Tests:
- Domain 10x10x10 + Halo (size: 0.2)
- Cutoff / cell length: 0.2
- Number of cells: 140608 (52x52x52)
- 1 million particles uniformly randomly distributed.
- Time in ms
- T = Number of threads

Description | Region    | Old (T=1) | New (T=1) | Old (T=8) | New (T=8) | Comment
------------ | ------------ | ------------ | ------------ | ------------ | ------------ | ------------ 
~6.4% of domain | [5 5 5] to [9 9 9]                     | 31 | 3 | 10 | 2 |probably realistic case
Full domain | [-0.2 -0.2 -0.2] to [10.2 10.2 10.2]  | 37 | 45 | 9 | 13 | Here the normal iterator should be used wich needs ~18 ms (T=1)
Full domain w/o halo | [0 0 0] to [10 10 10]         | 39 | 46 | 9 | 12 | Here the normal iterator should be used wich needs ~20 ms (T=1)
One halo side | [-0.2 -0.2 -0.2] to [10.2 10.2 0]   | 3 | <1 | 1 | <1 |
Half halo  | [-0.2 -0.2 -0.2] to [10.2 5 10.2]          | 3 | 2 | 1 | 1 |