@FG-TUM 2018-09-28
> what's the difference between:
> 
> > One halo side | [-0.2 -0.2 -0.2] to [10.2 10.2 0] | 3 | <1 |
> 
> and
> 
> > Half halo | [-0.2 -0.2 -0.2] to [10.2 10.2 0] | 3 | 2
> 
> ?

Typo :innocent: 
Half halo is until [10.2 **5** **10.2**]