@SteffenSeckler 2018-10-29
all ok, maybe comment, that the incSearchRegion should normally not be set manually.